tgbot/requirements.txt
aiogram[fast]
python-dotenv
uvicorn